here

+++ src/Http/Livewire/Journal/TempList.php
@@ -35,7 +35,9 @@ class TempList extends Component
         if (auth()->user()->can('2.delete.transfer.unapproved')) {
             Ledger::where('is_approve', 'f')->where('voucher_no', $voucher_no)->delete();
             LedgerAttachment::where('type', '0')->where('voucher_no', $voucher_no)->delete();
-            AmsCustomerPayment::where('voucher_no', $voucher_no)->where('temp_voucher', 't')->delete();
+            if(env('AMS_CUSTOMER', false) === true) {
+                AmsCustomerPayment::where('voucher_no', $voucher_no)->where('temp_voucher', 't')->delete();
+            }
             $this->success = 'Voucher #' . $voucher_no . ' has been deleted.';
         } else {
             $this->addError('success', env('PERMISSION_ERROR'));
@@ -64,10 +66,12 @@ class TempList extends Component
 
                 $ledger_entries = Ledger::where('is_approve', 'f')->where('voucher_no', $voucher_no)->select('id')->get();
 
-                AmsCustomerPayment::where('voucher_no', $voucher_no)->where('temp_voucher', 't')->update([
-                    'voucher_no' => $vno,
-                    'temp_voucher' => 'f',
-                ]);
+                if(env('AMS_CUSTOMER', false) === true) {
+                    AmsCustomerPayment::where('voucher_no', $voucher_no)->where('temp_voucher', 't')->update([
+                        'voucher_no' => $vno,
+                        'temp_voucher' => 'f',
+                    ]);
+                }
 
                 foreach ($ledger_entries as $e) {
                     Ledger::find($e->id)->update([
